*se trabajo en el guardar el spot *se desarrollo el viewpager para mostrar la informacion del spot , tiene bug *se desarrollo otra manera de capturar las imagenes y guardarlas en file del dispositivo , aun no se muestar ene le recycler

+++ app/src/main/java/go/goskate/goskate/vo/SpotVO.kt
@@ -12,8 +12,8 @@ class SpotVO {
     var hour = ""
     var files: List<File>? = null
     var score = 0
-    var latitude = 0F
-    var longitude = 0F
+    var latitude = 0.0
+    var longitude = 0.0
     var comments = ""
 
     constructor()
@@ -27,8 +27,8 @@ class SpotVO {
         hour: String,
         files: List<File>,
         score: Int,
-        latitude: Float,
-        longitude: Float,
+        latitude: Double,
+        longitude: Double,
         comments: String
     ) {
         this.nameSpot = nameSpot
